feat: add manage page
